quit on no input

+++ src/platform/Termbox.cc
@@ -706,10 +706,10 @@ drawTimeSlider()
     if (app::g_pMixer->bPaused) tb_set_cell(1, off, '|', TB_WHITE|TB_BOLD, TB_DEFAULT);
     else tb_set_cell(1, off, '>', TB_WHITE|TB_BOLD, TB_DEFAULT);
 
-    for (int i = xOff + 1; i < width - 1; ++i)
+    for (long i = xOff + 1; i < width - 1; ++i)
     {
         wchar_t wc = L'━';
-        if ((i - 1) == int(timePlace + xOff)) wc = L'╋';
+        if ((i - 1) == long(timePlace + xOff)) wc = L'╋';
         tb_set_cell(i, off, wc, TB_WHITE, TB_DEFAULT);
     }
 }
